update to Neo4j 2.2-SNAPSHOT

+++ runtime/src/main/java/com/graphaware/runtime/monitor/DatabaseLoadMonitor.java
@@ -10,5 +10,5 @@ public interface DatabaseLoadMonitor {
      *
      * @return load in tx per second.
      */
-    int getLoad();
+    long getLoad();
 }
